reusable eventlist component
