+++ timesketch/frontend-ng/src/views/Sketch.vue
@@ -30,7 +30,14 @@ limitations under the License.
         <v-icon>mdi-menu</v-icon>
       </v-btn>
 
-      <v-btn v-show="currentRouteName !== 'Explore'" :to="{ name: 'Explore', params: { sketchId: sketchId } }" color="primary" small text class="ml-3">
+      <v-btn
+        v-show="currentRouteName !== 'Explore'"
+        :to="{ name: 'Explore', params: { sketchId: sketchId } }"
+        color="primary"
+        small
+        text
+        class="ml-3"
+      >
         <v-icon small left>mdi-arrow-left</v-icon>
         back to explore
       </v-btn>
@@ -220,7 +227,7 @@ limitations under the License.
         <v-tab v-for="item in leftPanelTabItems" :key="item"> {{ item }} </v-tab>
       </v-tabs>
       <v-divider></v-divider>
-      <v-tabs-items v-model="leftPanelTab" class="pt-4">
+      <v-tabs-items v-model="leftPanelTab">
         <v-tab-item :transition="false">
           <ts-saved-searches v-if="meta.views"></ts-saved-searches>
           <ts-data-types></ts-data-types>
@@ -291,7 +298,7 @@ export default {
     return {
       showSketchMetadata: false,
       navigationDrawer: {
-        width: 400,
+        width: 370,
       },
       selectedScenario: null,
       scenarioDialog: false,
